Rework interface to MATLAB module.
The initialization now takes a MATLAB command instead of the MATLAB
root.  This uses the standard Boost.Build tools to find the MATLAB
program.

This has several advantages.  First, the PATH is searched, so if
MATLAB is on the path, the module auto-configures itself.  The version
is also queried and a particular version can be required.